Add support for mounting and unmounting devices
This commit adds support for mounting and unmounting storage devices.
The `mountmenu` script now takes one argument, either `mount` or `unmount`.
If `mount` is passed, the script will show a list of available devices
that can be mounted.
Once a device is selected, the user can choose where to mount it.
The script will create the mount point directory if it does not already exist
and then mount the device.

If `unmount` is passed,
the script will show a list of currently mounted devices.
Once a device is selected, it will be unmounted.

This functionality was implemented by adding a `launcher` function
that calls the `bemenu` menu.
The user can then select a device from the menu
and perform the mount or unmount operation.

The `Device` class was modified to remove the `model` attribute and to convert
the `size` attribute to an integer.
The `Table` class was modified to add a `is_empty` method to check
if there are any devices in the table and a `get_device_by_path`
method to retrieve a device given its path.
Additionally, the `__str__` method of the `Table` class was modified to display
the mount points of each device.