feat: Add ability to detect if a device is encrypted and display it in the output
The `Device` class now has an `is_encrypted` attribute
that checks if the device's filesystem type is either `crypto_LUKS` or `crypt`.
The `__str__` method of the `Device` class
has been updated to display this new attribute.

Additionally, when mounting a device, if it is encrypted, the user is prompted
to enter a name for the decrypted drive.
When unmounting an encrypted device, the `sudo cryptsetup close` command is also
executed on the device.

These changes improve the usability of the `mountmenu` script by providing more
information about the devices and adding support for encrypted devices.