feat: Add toggle mute and change volume functionality
This commit adds the `toggle_mute()` and `change_volume()` functions, which allow the user to toggle mute and adjust volume respectively. The script checks for the availability of the `wpctl` and `pulsemixer` commands to determine which function to use. If `wpctl` is available, it will be used to set volume and toggle mute. If not, `pulsemixer` will be used instead.

The `notify-send` command in `notify-music` has been updated to include the `--app-name=music` option to specify the name of the application sending the notification.

Additionally, the `toggle-mute` script now includes a `get_volume()` function to retrieve the current audio volume, an `is_muted()` function to check whether the audio is currently muted, and a call to `dunstify` to show a notification with the current volume and mute status.