Add change-brightness script and update gitignore and README
This commit adds a new script called "change-brightness" that can be used
to change the screen brightness and send notifications. The script uses
the 'brightnessctl' command to change the brightness level and 'libnotify'
to send notifications about the changed brightness level.

The 'README.md' file has been updated to reflect these changes and now lists
the 'change-brightness' script along with its dependencies.