Add usage instructions and notification option to change-brightness and change-volume scripts
Update the change-brightness and change-volume scripts to include a usage function that displays instructions on how to use the scripts, including examples.
Additionally, add an option to both scripts that allows users to display a desktop notification when adjusting the brightness or volume using the scripts.
Finally, update sb-brightness to include an option to open the change-brightness script when the left mouse button is clicked.