feat(counter): Add usage function, help option, and no-notification option to counter script
This commit adds a `usage` function, a `--help` option, and a `--no-notification` option to the `counter` script.
The `usage` function prints out a summary of how to use the script, including information about its options and positional arguments.
The `--help` option displays this usage summary when passed to the script.

The `--no-notification` option allows users to disable the desktop notification that is displayed when the countdown timer reaches zero. By default, the notification is enabled.

Additionally, this commit fixes a minor style issue by changing the shebang line from `#!/usr/bin/env sh` to `#!/bin/sh`.