feat(scratchpad): Add command line options and debug mode to scratchpad
This commit adds command line options and a debug mode to the `scratchpad` program. The program now accepts various options that allow users to customize the terminal emulator, working directory, window size, and more. Users can also enable the new `--debug` option to display debugging information about the program's internal state.

The new options are implemented using `getopt` and are parsed in a new loop at the beginning of the program. The options are then used to set various variables that control the behavior of the program.

The new `debug` function is added to display the values of the variables after they have been parsed from the command line options. This can be useful for debugging issues with the program's configuration.

The `scratchpad_terminal` variable has been renamed to `terminal` for clarity.

Overall, this commit makes `scratchpad` more flexible and easier to use by adding command line options and a debug mode.