refactor: Modify sleep and mpvpaper handling in wallpaperLoop script
This commit refactors the wallpaperLoop script by making changes to the sleep and mpvpaper handling. The following modifications have been made:

1. In the `kill_sleep` function, I added a check for an empty `$SLEEP_PID` variable. If it is empty, an error message is logged. Otherwise, the `SLEEP_PID` is killed and set to empty.

2. In the `quit_mpvpaper` function, I added a check for an empty `$MPVPAPER_PID` variable. If it is empty, an error message is logged. Otherwise, the `MPVPAPER_PID` is terminated.

3. In the `pause_sleep` function, I added a check for an empty `$SLEEP_PID` variable. If it is empty, an error message is logged. Otherwise, the `SLEEP_PID` is stopped.

4. In the `resume_sleep` function, I added a check for an empty `$SLEEP_PID` variable. If it is empty, an error message is logged. Otherwise, the `SLEEP_PID` is continued.

Additionally, the volume level for the MPV player has been changed from 10 to 24 by modifying the `mpv_options` variable.

These changes improve the handling of sleep and mpvpaper in the wallpaperLoop script for better functionality and error reporting.